substantial changes to the subroutine CN-allocation.c
Compared to the previous version, the fluxes explicitly declared as fluxes of CWD to the soil have been removed.
In the current version, the allocation takes in to account the newly assimilated carbon to be allocated
in the different tree pools, and the removal due to : leaf/froot/fruit senescence,or defoliation and the removal due to pruning. At this stage, no mortality of the entire tree is considered.