update on the output format
In the case no more forest class are in the stand (no living trees in the cell)
the simulation runs until the end of the expected temporal window.
The code print the output at cell level and soil level, only the variables strictly speaking
related to plants, are set to 0.