added new variable: tree to be removed in self thinning
the self thinning function is activated according to the density
of the layer, DBH of the class considered and the shade tollerance degree of the species.
The new added variable, indicate how many trees should be removed from the layer, in order
to have a DBHDC = DBHDC_min. This variable is used if the computed DBHDC value is lower than the minimum.