new function self thinning and modified self-pruning
SELF THINNING: in the new function, a prescribed number of trees (commputed in canopy_cover.c)
is removed from the lower forest class of the considered layer. This allows the larger DBH-class
(and theoretically, higher class in the layer) to suppress smaller trees and reach a DBHDC = DBHDC_min.
If the forest class in the lower height, has a number of trees lower than the number to the removed, the class is entirely removed. In the current version, the excess of trees left to be removed, is not considered.
In this case, it should be expected the activation of the self-thinning in the following year.

SELF PRUNING: pools of branch and croot and cell-level are updated directly in this function
(before, the branch and croot fraction to be removed, was considere in the CN_allocation.c).
 This avoid numerical issue when in the same day there is self pruning and self thinning.

AGE/STOCASTIC/ mortality: the number of trees to be removed is declared explicitly using the auxiliary
variable deadtree.