small change when no trees at cell level
If there are no trees, because of harvest, in the output file, it is reported
the total amount of carbon removed as HWP on the first day of the year.
However, as everything has been removed, no data for tress, e.g. GPP, NPP etc,
are reported and species is indicated as "none"