save mortality data in the output file
When natural mortality occurs and an entire forest class is removed,
data are printed in the output file.
If only one percentage of the trees are removed, this datum is updated
in the variable DEAD_TREE and this information is used at the end of the year
in the file tree_daily to print.